requirements.txt
apache-airflow-providers-postgres==4.0.1
apache-airflow-providers-http==2.1.1
pandas==1.4.1